requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
joblib>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
ta>=0.10.0
yfinance>=0.2.0
plotly>=5.0.0
dash>=2.10.0
pytest>=7.0.0
jupyter>=1.0.0
python-binance>=1.0.0
python-dotenv>=0.20.0

xgboost
optuna

pymc

scipy

matplotlib
